requirements.txt
mpmath==1.3.0
numpy==2.3.1
pillow==11.3.0
pyperclip==1.9.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
sympy==1.14.0
